python/NagyZH2/requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-migrate
